Moved asteroids generation to separate file.
